Preliminary fix for zombie <Defunct> HAProxy processes
Python 2 does not automatically clean up zombie processes. An issue was discovered during Bakerscale testing where the OS process list became clogged with zombie HAProxies due to not waiting for the process to finish. Calling wait() solves this problem. In the future an alternative approach might be to append processes into a list and run a thread that cleans them without waiting for the HAProxy process reload to complete.